requirements.txt
Pillow==5.4.1
matplotlib
numpy
pypylon==1.3.1
pyueye==4.90.0.0
opencv-python